Revert "move project to typescript"
This reverts commit ea126e5790d3ca7adf9db83f99420f8412574f8f.
