+++ .github/actions/surge-preview-tools/.gitignore
@@ -1,2 +1,1 @@
-lib/
 node_modules/
