+++ .github/actions/surge-preview-tools/package.json
@@ -4,16 +4,14 @@
   "license": "Apache-2.0",
   "description": "",
   "scripts": {
-    "build": "tsc && ncc build"
+    "build": "ncc build index.js --license licenses.txt"
   },
   "dependencies": {
     "@actions/core": "~1.9.0",
     "@actions/github": "~5.0.3",
-    "strip-ansi": "~7.0.1"
+    "strip-ansi": "~6.0.1"
   },
   "devDependencies": {
-    "@types/node": "~16.11.7",
-    "@vercel/ncc": "~0.34.0",
-    "typescript": "~4.7.3"
+    "@vercel/ncc": "~0.34.0"
   }
 }
