--- .github/actions/surge-preview-tools/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es6",
-    "module": "commonjs",
-    "moduleResolution": "Node",
-    "outDir": "./lib",
-    "rootDir": "./src",
-    "declaration": false,
-    "strict": true,
-    "noImplicitAny": true,
-    "esModuleInterop": true,
-  },
-  "exclude": ["node_modules", "**/*.test.ts"],
-  "include": ["src"]
-}
